requirements.txt
PyPDF2
langchain
sentence-transformers
chromadb
gradio
python-dotenv
langchain_groq
langchain-community
streamlit
langchain-google-genai
pysqlite3-binary